Add AudioTranscriptionFile enum (#62)
* Add AudioTranscriptionFile enum

This is in preparation of allowing to pass the audio file content as
bytes rather than a filename on the local filesystem.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>

* Add Bytes variant to AudioTranscriptionFile

This variant makes it possible to pass the audio file content as bytes,
making the library more flexible. E.g. it now becomes possible to get
the audio file via a network connection and pass it directly without
having to use a temporary file.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>

---------

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>